+++ james-controller-consul/src/main/java/com/tomtom/james/controller/consul/dto/InformationPointDTOv1.java
@@ -24,8 +24,8 @@ public class InformationPointDTOv1 extends Versioned {
 
     private List<String> script;
     private Integer sampleRate;
-    private Integer successSampleRate;
-    private Integer errorSampleRate;
+    private Double successSampleRate;
+    private Double errorSampleRate;
     private Metadata metadata;
 
     public List<String> getScript() {
@@ -36,11 +36,11 @@ public class InformationPointDTOv1 extends Versioned {
         return sampleRate;
     }
 
-    public Integer getSuccessSampleRate() {
+    public Double getSuccessSampleRate() {
         return successSampleRate;
     }
 
-    public Integer getErrorSampleRate() {
+    public Double getErrorSampleRate() {
         return errorSampleRate;
     }
